refactor downloading into a separate Downloader class
